[Project] Fix `-Wconversion` warnings for Qt6
This commit introduces new functions `qsizetype_to_int` and
`safe_int_cast` for integer conversion.  In Qt6, the type of `qsizetype`
changes from `int` to `long long int`.  And functions such as `size()`
no longer return `int`.  This resulted in hundreds of conversion
warnings for Qt6.  With this commit, `qsizetype_to_int` does a checked
conversion for Qt6, and simply returns the int in Qt5.

`elch_size_t` is used as `qsizetype` was introduced in Qt 5.10, but we
have to support earlier versions as well.

While most checks are probably not needed (e.g. we don't use more than
2^31 actors), they don't impact performance too much.  Better be safe
than sorry.

+++ src/ui/main/QuickOpen.cpp
@@ -194,7 +194,7 @@ void QuickOpen::updateViewGeometry()
     const QSize centralSize = parentWidget()->size();
 
     // width: 2.4 of editor, height: 1/2 of editor
-    const QSize viewMaxSize(centralSize.width() / 2.4, centralSize.height() / 2);
+    const QSize viewMaxSize(static_cast<int>(centralSize.width() / 2.4), static_cast<int>(centralSize.height() / 2.));
 
     // Position should be central over window
     const int xPos = std::max(0, (centralSize.width() - viewMaxSize.width()) / 2);
